Issue #1110: Added capability to remove reference data items
